+++ matlab/source/cepton_sdk_matlab.h
@@ -1,9 +1,9 @@
 #ifndef CEPTON_SDK_MATLAB_H
 #define CEPTON_SDK_MATLAB_H
 
-#define SIMPLE
+#define CEPTON_SIMPLE
 #include "cepton_sdk.h"
-#undef SIMPLE
+#undef CEPTON_SIMPLE
 
 #ifdef __cplusplus
 extern "C" {
@@ -12,34 +12,35 @@ extern "C" {
 #include "cepton_sdk_def.h"
 
 /*
- * Duplicating some functions to provide SIMPLE interface for MATLAB.
- * 
+ * Duplicating some functions to provide interface for MATLAB.
+ *
  * MATLAB doesn't support:
  * - Nested structs.
  * - Function pointers.
  * - Unions.
-*/
+ */
 
-EXPORT CeptonSensorErrorCode
+CEPTON_EXPORT CeptonSensorErrorCode
 cepton_sdk_matlab_initialize(int ver, CeptonSDKControl control_flags);
-EXPORT CeptonSensorErrorCode cepton_sdk_matlab_deinitialize();
-EXPORT CeptonSensorErrorCode cepton_sdk_matlab_clear_cache();
+CEPTON_EXPORT CeptonSensorErrorCode cepton_sdk_matlab_deinitialize();
+CEPTON_EXPORT CeptonSensorErrorCode cepton_sdk_matlab_clear_cache();
 
-EXPORT CeptonSensorErrorCode cepton_sdk_matlab_get_sensor_information(
+CEPTON_EXPORT CeptonSensorErrorCode cepton_sdk_matlab_get_sensor_information(
     CeptonSensorHandle sensor_handle,
     struct CeptonSensorInformation *const sensor_info);
-EXPORT CeptonSensorErrorCode cepton_sdk_matlab_get_sensor_information_by_index(
+CEPTON_EXPORT CeptonSensorErrorCode
+cepton_sdk_matlab_get_sensor_information_by_index(
     int sensor_index, struct CeptonSensorInformation *const sensor_info);
 
-EXPORT int cepton_sdk_matlab_has_error();
-EXPORT CeptonSensorErrorCode cepton_sdk_matlab_get_error(
+CEPTON_EXPORT int cepton_sdk_matlab_has_error();
+CEPTON_EXPORT CeptonSensorErrorCode cepton_sdk_matlab_get_error(
     CeptonSensorHandle *const handle, CeptonSensorErrorCode *const error_code,
     size_t *const msg_size);
 
-EXPORT int cepton_sdk_matlab_has_image_points();
-EXPORT CeptonSensorErrorCode cepton_sdk_matlab_get_image_points(
+CEPTON_EXPORT int cepton_sdk_matlab_has_image_points();
+CEPTON_EXPORT CeptonSensorErrorCode cepton_sdk_matlab_get_image_points(
     CeptonSensorHandle *const sensor_handle, size_t *const n_points);
-EXPORT CeptonSensorErrorCode cepton_sdk_matlab_get_image_points_data(
+CEPTON_EXPORT CeptonSensorErrorCode cepton_sdk_matlab_get_image_points_data(
     uint64_t *const timestamps, float *const image_x, float *const distances,
     float *const image_z, float *const intensities, uint8_t *const return_types,
     uint8_t *const flags);